реализовал удаление и лайк карточек

+++ src/utils/Api.js
@@ -62,20 +62,20 @@ class Api {
     }).then(this._checkQueryResult);
   }
 
-  setLikeCard(cardId) {
-    return fetch(`${this._baseUrl}/cards/${cardId}/likes`, {
-      method: "PUT",
-      headers: this._headers,
-    }).then(this._checkQueryResult);
-  }
-
-  deleteLikeCard(cardId) {
-    return fetch(`${this._baseUrl}/cards/${cardId}/likes`, {
-      method: "DELETE",
-      headers: this._headers,
-    }).then(this._checkQueryResult);
-  }
-}
+changeLikeCardStatus(cardId, isLiked) {
+    if(!isLiked) {
+      return fetch(`${this._baseUrl}/cards/${cardId}/likes`, {
+        method: "DELETE",
+        headers: this._headers,
+      }).then(this._checkQueryResult);
+    } else {
+      return fetch(`${this._baseUrl}/cards/${cardId}/likes`, {
+        method: "PUT",
+        headers: this._headers,
+      }).then(this._checkQueryResult);
+    }
+    }
+  } 
 
 const api = new Api({
   baseUrl: "https://mesto.nomoreparties.co/v1/cohort-71",
